requirements.txt
streamlit==1.43.1
torch==2.6.0
yolov5==7.0.14
pillow==11.1.0
numpy==2.1.1
opencv-python-headless==4.11.0.86